refactor: moved budget typedefs to common folder
