+++ client/src/components/ActiveBudget/ActiveBudget.types.ts
@@ -1,4 +1,4 @@
-import { IBudget } from '../../pages/Budget/Budget.types';
+import { IBudget } from '../../types/Budget.types';
 
 export interface IProps {
     monthBudget: IBudget
