+++ client/src/pages/Budget/Budget.types.ts
@@ -16,18 +16,3 @@ export interface IBudgetDatum {
     diffPc: number
     variance: [number, number]
 }
-
-export interface IBudgetRow {
-    label: string
-    value: number
-    varLowPc: number
-    varHighPc: number
-}
-
-export interface IBudget {
-    id: number
-    name: string
-    budget: {
-        [key: number]: IBudgetRow
-    }
-}
