+++ client/src/pages/Budget/components/GlanceCards/GlanceCards.types.ts
@@ -1,4 +1,6 @@
-import { IBudget, IBudgetDatum } from '../../Budget.types';
+import { IBudget } from '../../../../types/Budget.types';
+
+import { IBudgetDatum } from '../../Budget.types';
 
 export interface IProps {
     data: IBudgetDatum[]
