+++ client/src/types/Budget.types.ts
@@ -1,0 +1,14 @@
+export interface IBudgetRow {
+    label: string
+    value: number
+    varLowPc: number
+    varHighPc: number
+}
+
+export interface IBudget {
+    id: number
+    name: string
+    budget: {
+        [key: number]: IBudgetRow
+    }
+}
